+++ Cargo.toml
@@ -16,6 +16,7 @@ readme = "README.md"
 weresocool_parser = { path = "parser" }
 weresocool_ast = { path = "ast" }
 weresocool_error = { path = "error" } 
+weresocool_lame = { path = "lame" } 
 portaudio = "0.7.0"
 rand = "0.7.3"
 hound = "3.4.0"
@@ -41,13 +42,10 @@ bytes = "0.5.4"
 env_logger = "0.7.1"
 test-generator = "0.3.0"
 float-cmp = "0.8.0"
-lame = "0.1.3"
-lame-sys = "0.1.2"
 
 [dev-dependencies]
 pretty_assertions = "0.6.1"
 criterion = "0.3"
-#listenfd = "0.3.3"
 
 [[bench]]
 name = "render_batch_bench"
@@ -59,6 +57,7 @@ harness = false
 [workspace]
 members = [
     "ast",
+    "lame",
     "parser",
     "error",
     "application/server",
