+++ error/Cargo.toml
@@ -7,6 +7,7 @@ description = "Errors for WereSoCool"
 license = "GPL-3.0"
 
 [dependencies]
+weresocool_lame = { path = "../lame" }
 portaudio = "0.7.0"
 hound = "3.4.0"
 serde_json = "1.0.53"
