+++ error/src/error_inner.rs
@@ -27,4 +27,13 @@ pub enum ErrorInner {
 
     #[fail(display = "Index error: {}", _0)]
     IndexError(#[cause] IndexError),
+
+    #[fail(display = "Hound error: {}", _0)]
+    HoundError(#[cause] hound::Error),
+
+    #[fail(display = "Lame error: {}", _0)]
+    LameError(#[cause] weresocool_lame::Error),
+
+    #[fail(display = "LameEncode error: {}", _0)]
+    LameEncodeError(#[cause] weresocool_lame::EncodeError),
 }
