+++ src/generation/mod.rs
@@ -1,6 +1,4 @@
-pub mod ffi;
 pub mod json;
-pub mod lame;
 pub mod parsed_to_render;
 
 pub mod test;
